requirements.txt
torch==2.4.1
langchain==0.2.16
langchain-community==0.2.16
langchain-groq==0.1.9
langchain-chroma==0.1.2
langchain-huggingface==0.0.3
langchain-experimental==0.0.65
sentence-transformers==3.0.1
chromadb==0.5.5
pymupdf==1.24.9
docx2txt==0.8
nemoguardrails==0.10.1
streamlit==1.38.0
python-dotenv==1.0.1
groq==0.9.0
httpx==0.27.2
httpcore==1.0.5
